+++ sdk/nodejs/tsconfig.json
@@ -47,6 +47,7 @@
 
         "cmd/dynamic-provider/index.ts",
         "cmd/run/index.ts",
+        "cmd/run/run.ts",
 
         "tests/config.spec.ts",
         "tests/init.spec.ts",
